feat(Name, Location): 자동차의 위치와 이름을 포장하는 클래스 생성 refactor(Car): Car의 불변 객체로 변경
