+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,33 +1,29 @@
 package racingcar.domain.car;
 
 import racingcar.domain.car.strategy.MovementStrategy;
+import racingcar.vo.car.Location;
+import racingcar.vo.car.Name;
 
 import java.util.Objects;
 
-public final class Car {
-    private final String name;
-    private int location;
+public class Car {
+    private final Name name;
+    private final Location location;
 
     public Car(String name) {
-        validateName(name);
-        this.name = name;
+        this.name = new Name(name);
+        this.location = new Location();
     }
 
     public int move(MovementStrategy movementStrategy) {
         if (movementStrategy.isMovable()) {
-            location++;
+            location.forward();
         }
 
-        return location;
+        return location.getPosition();
     }
 
     public String getName() {
-        return name;
-    }
-
-    private void validateName(String name) {
-        if (Objects.isNull(name) || name.trim().isEmpty()) {
-            throw new IllegalArgumentException("이름은 Null 또는 공백이 될 수 없습니다.");
-        }
+        return name.getName();
     }
 }
