+++ src/main/java/racingcar/vo/car/Location.java
@@ -1,0 +1,13 @@
+package racingcar.vo.car;
+
+public class Location {
+    private int position;
+
+    public void forward() {
+        this.position++;
+    }
+
+    public int getPosition() {
+        return this.position;
+    }
+}
