+++ src/main/java/racingcar/vo/car/Name.java
@@ -1,0 +1,22 @@
+package racingcar.vo.car;
+
+import java.util.Objects;
+
+public class Name {
+    private final String name;
+
+    public Name(String name) {
+        validateName(name);
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    private void validateName(String name) {
+        if (Objects.isNull(name) || name.trim().isEmpty()) {
+            throw new IllegalArgumentException("이름은 Null 또는 공백이 될 수 없습니다.");
+        }
+    }
+}
